updating readme, renaming file
