+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.4.2 (October 17, 2020)
 
-- Warn if a config is undefined
+- Warn if a config is undefined, null, 'undefined' or an empty string
 
 ## 1.4.1 (October 17, 2020)
