+++ src/validate-config.ts
